Fix for new testcase: nested textile tags producing broken HTML tags
Discovered when encountering \*ABC\* which was converted to:

    <strong>span class="caps">ABC</span></strong>

Issue was found in the lex_attrs method that looked for alignment
indicators. It discovers html tags from previous replacements and
considers them to be left alignment operators.

Added a rough regex to spot an opening HTML tag to prevent this from
happening

---

Added a test case and minified.